Maintain precision

+++ src/cis-ase/beta_model.h
@@ -106,21 +106,21 @@ class BetaModel {
         }
         //Calculate likelihood under the S model
         void calc_S_lik() {
-            float AF = alt_count_/(ref_count_ + alt_count_);
+            float AF = (float)alt_count_/(float)(ref_count_ + alt_count_);
             bool log_density = false;
             lik_S_ = 0.5 * (dbeta(AF, S_alpha, S_beta, log_density) +
                            dbeta(AF, S_beta, S_alpha, log_density));
         }
         //Calculate likelihood under the M model
         void calc_M_lik() {
-            float AF = alt_count_/(ref_count_ + alt_count_);
+            float AF = (float)alt_count_/(float)(ref_count_ + alt_count_);
             bool log_density = false;
             lik_M_ = 0.5 * (dbeta(AF, M_alpha, M_beta, log_density) +
                            dbeta(AF, M_beta, M_alpha, log_density));
         }
         //Calculate likelihood under the N model
         void calc_N_lik() {
-            float AF = alt_count_/(ref_count_ + alt_count_);
+            float AF = (float)alt_count_/(float)(ref_count_ + alt_count_);
             bool log_density = false;
             lik_N_ = 0.5 * (dbeta(AF, N_alpha, N_beta, log_density) +
                            dbeta(AF, N_beta, N_alpha, log_density));
